Remove remaining int usage and update docs

+++ src/equations/summary.c
@@ -16,7 +16,7 @@ void equations_summary(const Equations *eqns)
     long num = eqns->boundary.num;
     Name *name = eqns->boundary.name;
 
-    int width = 0;
+    long width = 0;
     for (long i = num_inner; i < num_entities; i++) {
         width = lmax(width, strlen(entity[i]));
     }
